pyproject.toml
[project]
name = "tribute120"
version = "0.1.0"
description = ""
authors = [
    {name = "favouritetasty",email = "19307110112@fudan.edu.cn"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
